入院患者数: Rename components/cards/HospitalizedNumberCard.vue to components/index/CardsMonitoring/HospitalizedNumber/Card.vue
