+++ components/index/CardsMonitoring/HospitalizedNumber/Card.vue
