+++ libraries/repositories/PositiveStatusRepository.ts
@@ -19,7 +19,7 @@ export class PositiveStatusRepository
   /**
    * 使用箇所
    *
-   * モニタリング項目(6) 入院患者数 (components/cards/HospitalizedNumberCard.vue)
+   * モニタリング項目(6) 入院患者数 (components/index/CardsMonitoring/HospitalizedNumber/Card.vue)
    *
    * モニタリング項目(7) 重症患者数 (components/cards/SevereCaseCard.vue)
    */
